// frontend/src/api/generated/newChatbotAPI.schemas.ts
export const GetAdminCouponsId200AllOfTwoDataDiscountType = {  PERCENTAGE: 'PERCENTAGE',
  FIXED: 'FIXED',
} as const
export type GetAdminCouponsId200AllOfTwoData = {
  id: string;
  code: string;
  discountType: typeof GetAdminCouponsId200AllOfTwoDataDiscountType[keyof typeof GetAdminCouponsId200AllOfTwoDataDiscountType] ;
  value: number;
  expiresAt?: string;
  usageLimit?: number;
  usageCount: number;
  active: boolean;
  stripeCouponId?: string;
  createdAt: string;
  updatedAt: string;
};